Release nxtvepg-2.5.2 (2003-Jul-01)
Changes between 2.5.1 and 2.5.2 (as per file CHANGES):
- Windows BT8x8 driver: restore settings for all PCI registers upon shutdown
  (solves problem with Hauppauge WinTV, thanks to Peter Rolf for pointing this
  out; also solves problem with BtWinCap driver, as pointed out anonymously
  in the discussion forum 2003-June-07)
- Windows: fixed support for Microtune tuner "MT2032" (e.g. on Pinnacle cards)
  Many thanks to Oliview Avenel for debugging support.
- Windows: updated & added TV card configuration from DScaler CVS
- Windows: added "Close this menu" entry to systray menu (workaround)
- removed German network 3sat from list of known Nextview EPG providers
- XML export via control menu dialog no longer applies current filter setting,
  i.e. all programmes of all networks in the database are exported.
- fixed compiler warnings for gcc 3.3 (signed/unsigned type conflicts)
- UNIX: added compiler switch -DZVBI_DECODER in the Makefile to replace VBI
  (Teletext) slicer with code from libzvbi which performs better on channels
  with weak reception; sometimes this slicer seems to eat up a lot of CPU
  (up to 25% on my 400MHz K6-2)

See file CHANGES.RCS for a detailed log of revisions of all files.

+++ epgdb/epgdbmgmt.c
@@ -21,7 +21,7 @@
  *
  *  Author: Tom Zoerner
  *
- *  $Id: epgdbmgmt.c,v 1.44 2003/01/11 19:40:20 tom Exp tom $
+ *  $Id: epgdbmgmt.c,v 1.45 2003/06/28 10:48:12 tom Exp tom $
  */
 
 #define DEBUG_SWITCH DEBUG_SWITCH_EPGDB
@@ -613,15 +613,18 @@ static bool EpgDbGenericBlockNoValid( PDBC dbc, EPGDB_BLOCK * pBlock, BLOCK_TYPE
       {
          case BLOCK_TYPE_NI:
             // block numbers of NI blocks start with 1, hence <=
-            accept = (block_no <= (dbc->pAiBlock->blk.ai.niCount + dbc->pAiBlock->blk.ai.niCountSwo));
+            accept = (block_no <= ((uint)dbc->pAiBlock->blk.ai.niCount +
+                                   (uint)dbc->pAiBlock->blk.ai.niCountSwo));
             break;
 
          case BLOCK_TYPE_OI:
-            accept = (block_no < (dbc->pAiBlock->blk.ai.oiCount + dbc->pAiBlock->blk.ai.oiCountSwo));
+            accept = (block_no < ((uint)dbc->pAiBlock->blk.ai.oiCount +
+                                  (uint)dbc->pAiBlock->blk.ai.oiCountSwo));
             break;
 
          case BLOCK_TYPE_MI:
-            accept = (block_no < (dbc->pAiBlock->blk.ai.miCount + dbc->pAiBlock->blk.ai.miCountSwo));
+            accept = (block_no < ((uint)dbc->pAiBlock->blk.ai.miCount +
+                                  (uint)dbc->pAiBlock->blk.ai.miCountSwo));
             break;
 
          case BLOCK_TYPE_LI:
